% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.initials.R
\name{get.initials}
\alias{get.initials}
\title{Get initial values for spexvb}
\usage{
get.initials(
  X,
  Y,
  mu = NULL,
  omega = NULL,
  c_pi = NULL,
  d_pi = NULL,
  intercept = FALSE,
  tau_e = NULL,
  update_order = NULL,
  seed = 12376
)
}
\arguments{
\item{X}{A design matrix.}

\item{Y}{A response vector.}

\item{mu}{Initial mean.}

\item{omega}{Initial omega.}

\item{c_pi}{Initial c_pi.}

\item{d_pi}{Initial d_pi.}

\item{intercept}{Logical, whether an intercept is included.}

\item{tau_e}{Initial tau_e.}

\item{update_order}{Initial update order.}

\item{seed}{Seed for reproducibility.}
}
\value{
A list of initialized parameters.
}
\description{
This function initializes parameters for the spexvb model.
}
\details{
Generate Initial Values for Variational Inference in Sparse Regression

This helper function estimates initial values for variational parameters such as
regression coefficients (\code{mu}), spike probabilities (\code{omega}), and hyperparameters
like \code{tau_e}, \code{c_pi}, and \code{d_pi} using LASSO and Ridge regression fits.
}
