% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_initials_spxlvb.R
\name{get_initials_spxlvb}
\alias{get_initials_spxlvb}
\title{Get initial values for spxlvb}
\usage{
get_initials_spxlvb(
  X,
  Y,
  mu_0 = NULL,
  omega_0 = NULL,
  c_pi_0 = NULL,
  d_pi_0 = NULL,
  tau_e = NULL,
  update_order = NULL,
  seed = 12376
)
}
\arguments{
\item{X}{A design matrix.}

\item{Y}{A response vector.}

\item{mu_0}{Initial mean.}

\item{omega_0}{Initial omega.}

\item{c_pi_0}{Initial c_pi.}

\item{d_pi_0}{Initial d_pi.}

\item{tau_e}{Initial tau_e.}

\item{update_order}{Initial update order.}

\item{seed}{Seed for reproducibility.}
}
\value{
A named list with elements \code{mu_0}, \code{omega_0},
\code{c_pi_0}, \code{d_pi_0}, \code{tau_e}, and \code{update_order}.
}
\description{
This function initializes parameters for the spxlvb model.
}
\details{
Generate Initial Values for Variational Inference in Sparse Regression

This helper function estimates initial values for variational parameters such as
regression coefficients (\code{mu}), spike probabilities (\code{omega}), and hyperparameters
like \code{tau_e}, \code{c_pi}, and \code{d_pi} using LASSO and Ridge regression fits.
}
\examples{
\donttest{
set.seed(1)
X <- matrix(rnorm(30 * 20), 30, 20)
Y <- rnorm(30)
init <- get_initials_spxlvb(X = X, Y = Y)
}
}
