% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mattern_data.R
\name{matern.data.gen}
\alias{matern.data.gen}
\title{Simulate Matern Correlated High-Dimensional Data}
\usage{
matern.data.gen(
  seed_val,
  n,
  p,
  pi_0,
  L = NULL,
  sd_beta = 1,
  SNR = 2,
  n_test = NULL,
  n_validation = NULL
)
}
\arguments{
\item{seed_val}{An integer. The random seed for reproducibility.}

\item{n}{An integer. The number of observations.}

\item{p}{An integer. The number of predictors.}

\item{pi_0}{A numeric value between 0 and 1. The proportion of non-zero coefficients.}

\item{L}{The Lower Cholesky matrix provided by \code{get.L}.}

\item{sd_beta}{A numeric value. The standard deviation of the non-zero coefficients.}

\item{SNR}{A numeric value. The signal-to-noise ratio.}

\item{n_test}{An integer. Number of test observations.}

\item{n_validation}{An integer. Number of validation observations.}
}
\description{
This function generates a synthetic dataset where the predictors (X)
follow a spatial correlation structure defined by a pre-computed L matrix.
}
