% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mattern_data.R
\name{get.L}
\alias{get.L}
\title{Compute Cholesky Factor for Matern Correlation}
\usage{
get.L(p, range, smoothness, phi = 1)
}
\arguments{
\item{p}{An integer. The number of predictors (dimension).}

\item{range}{A numeric value. Controls the distance at which correlation decays.}

\item{smoothness}{A numeric value (nu). Controls differentiability (0.5, 1.5, or 2.5).}

\item{phi}{A numeric value. The marginal variance. Defaults to 1.0.}
}
\description{
Pre-computes the Lower Cholesky factor (L) for a 1D Matern
covariance matrix. This should be called once per simulation setting to
save computational time.
}
